#day4
 Python List Methods:
 list index(),list append(),list count(),list sort(),list extend(),list insert(),list remove(),list pop(),list reverse(),list copy(),list clear()

1. list index()
The syntax of the list index() method is:

list.index(element, start, end)

list index() parameters
The list index() method can take a maximum of three arguments:

element - the element to be searched
start (optional) - start searching from this index
end (optional) - search the element up to this index

Return Value from List index()
The index() method returns the index of the given element in the list.
If the element is not found, a ValueError exception is raised.
Note: The index() method only returns the first occurrence of the matching element.
e.g.Index of the Element not Present in the List
# vowels list
vowels = ['a', 'e', 'i', 'o', 'u']

# index of'p' is vowels
index = vowels.index('p')
print('The index of p:', index)

output: ValueError: 'p' is not in list

2.list append()
The syntax of the append() method is:

list.append(item)
append() Parameters
The method takes a single argument

item - an item to be added at the end of the list
The item can be numbers, strings, dictionaries, another list, and so on.

Return Value from append()
The method doesn't return any value (returns None).

3.list count()
The syntax of the count() method is:

list.count(element)
count() Parameters
The count() method takes a single argument:

element - the element to be counted
Return value from count()
The count() method returns the number of times element appears in the list.

4.list sort()
The syntax of the sort() method is:

list.sort(key=..., reverse=...)
Alternatively, you can also use Python's built-in sorted() function for the same purpose.

sorted(list, key=..., reverse=...)
Note: The simplest difference between sort() and sorted() is: sort() changes the list directly and doesn't return any value, while sorted() doesn't change the list and returns the sorted list.

sort() Parameters
By default, sort() doesn't require any extra parameters. However, it has two optional parameters:

reverse - If True, the sorted list is reversed (or sorted in Descending order)
key - function that serves as a key for the sort comparison
Return value from sort()
The sort() method doesn't return any value. Rather, it changes the original list.

If you want a function to return the sorted list rather than change the original list, use sorted().

Sort with custom function using key
If you want your own implementation for sorting, the sort() method also accepts a key function as an optional parameter.

Based on the results of the key function, you can sort the given list.

list.sort(key=len)
Alternatively for sorted:

sorted(list, key=len)

Here, len is the Python's in-built function to count the length of an element.

The list is sorted based on the length of each element, from lowest count to highest.

We know that a tuple is sorted using its first parameter by default. Let's look at how to customize the sort() method to sort using the second element.

5.list extend()
The syntax of the extend() method is:

list1.extend(iterable)
Here, all the elements of iterable are added to the end of list1.

extend() Parameters
As mentioned, the extend() method takes an iterable such as list, tuple, string etc.

Return Value from extend()
The extend() method modifies the original list. It doesn't return any value.

*Python extend() Vs append()
If you need to add an element to the end of a list, you can use the append() method.
e.g.
a1 = [1, 2]
a2 = [1, 2]
b = (3, 4)

# a1 = [1, 2, 3, 4]
a1.extend(b) 
print(a1)

# a2 = [1, 2, (3, 4)]
a2.append(b)
print(a2)

Output
[1, 2, 3, 4]
[1, 2, (3, 4)]

6.list insert()
The list insert() method inserts an element to the list at the specified index.

The syntax of the insert() method is

list.insert(i, elem)
Here, elem is inserted to the list at the ith index. All the elements after elem are shifted to the right.

insert() Parameters
The insert() method takes two parameters:

index - the index where the element needs to be inserted
element - this is the element to be inserted in the list
Notes:

If index is 0, the element is inserted at the beginning of the list.
If index is 3, the element is inserted after the 3rd element. Its position will be 4th.
Return Value from insert()
The insert() method doesn't return anything; returns None. It only updates the current list.

7.list remove()
The remove() method removes the first matching element (which is passed as an argument) from the list.

The syntax of the remove() method is:

list.remove(element)
remove() Parameters
The remove() method takes a single element as an argument and removes it from the list.
If the element doesn't exist, it throws ValueError: list.remove(x): x not in list exception.
e.g.Deleting element that doesn't exist
# animals list
animals = ['cat', 'dog', 'rabbit', 'guinea pig']

# Deleting 'fish' element
animals.remove('fish')

# Updated animals List
print('Updated animals list: ', animals)

Output

Traceback (most recent call last):
  File ".. .. ..", line 5, in <module>
    animal.remove('fish')
ValueError: list.remove(x): x not in list
Return Value from remove()
The remove() doesn't return any value (returns None).

If you need to delete elements based on the index (like the fourth element), you can use the pop() method.
Also, you can use the Python del statement to remove items from the list.

8.list pop()
The pop() method removes the item at the given index from the list and returns the removed item.

The syntax of the pop() method is:

list.pop(index)
pop() parameters
The pop() method takes a single argument (index).
The argument passed to the method is optional. If not passed, the default index -1 is passed as an argument (index of the last item).
If the index passed to the method is not in range, it throws IndexError: pop index out of range exception.
Return Value from pop()
The pop() method returns the item present at the given index. This item is also removed from the list.

Note: Index in Python starts from 0, not 1.

If you need to pop the 4th element, you need to pass 3 to the pop() method.

9.list reverse()
The reverse() method reverses the elements of the list.

The syntax of the reverse() method is:

list.reverse()
reverse() parameter
The reverse() method doesn't take any arguments.

Return Value from reverse()
The reverse() method doesn't return any value. It updates the existing list.

10.list copy()
The copy() method returns a shallow copy of the list.

A list can be copied using the = operator. For example,

old_list = [1, 2, 3]
​new_list = old_list
The problem with copying lists in this way is that if you modify new_list, old_list is also modified.

old_list = [1, 2, 3]
new_list = old_list

# add an element to list
new_list.append('a')

print('New List:', new_list)
print('Old List:', old_list)

Output

Old List: [1, 2, 3, 'a']
New List: [1, 2, 3, 'a']
However, if you need the original list unchanged when the new list is modified, you can use the copy() method.

The syntax of the copy() method is:

new_list = list.copy()
copy() parameters
The copy() method doesn't take any parameters.

Return Value from copy()
The copy() method returns a new list. It doesn't modify the original list.

11.list clear()
The clear() method removes all items from the list.

The syntax of clear() method is:

list.clear()
clear() Parameters
The clear() method doesn't take any parameters.

Return Value from clear()
The clear() method only empties the given list. It doesn't return any value

Note: If you are using Python 2 or Python 3.2 and below, you cannot use the clear() method. You can use the del operator instead.